Add tests to SceneContainerComponentTests.
