+++ SceneryWebApi/Scenery.ComponentsTests/SceneComponents/CylinderSceneComponentTests.cs
@@ -1,9 +1,9 @@
-﻿// <copyright file="ConeSceneComponentTests.cs" company="dsnouck">
+﻿// <copyright file="CylinderSceneComponentTests.cs" company="dsnouck">
 // Copyright (c) dsnouck. All rights reserved.
 // Licensed under the MIT License. See LICENSE file in the project root for full license information.
 // </copyright>
 
-namespace Scenery.Components.Tests.SceneComponentTests
+namespace Scenery.Components.Tests.SceneComponents
 {
     using System;
     using System.Collections.Generic;
@@ -16,31 +16,31 @@ namespace Scenery.Components.Tests.SceneComponentTests
     using Xunit;
 
     /// <summary>
-    /// Provides tests for <see cref="ConeSceneComponent"/>.
+    /// Provides tests for <see cref="CylinderSceneComponent"/>.
     /// </summary>
-    public class ConeSceneComponentTests
+    public class CylinderSceneComponentTests
     {
-        private readonly ConeSceneComponent systemUnderTest;
+        private readonly CylinderSceneComponent systemUnderTest;
         private readonly Mock<IFuncDoubleDoubleComponent> funcDoubleDoubleComponentTestDouble;
         private readonly Mock<ILine3Component> line3ComponentTestDouble;
         private readonly Mock<IVector3Component> vector3ComponentTestDouble;
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="ConeSceneComponentTests"/> class.
+        /// Initializes a new instance of the <see cref="CylinderSceneComponentTests"/> class.
         /// </summary>
-        public ConeSceneComponentTests()
+        public CylinderSceneComponentTests()
         {
             this.funcDoubleDoubleComponentTestDouble = new Mock<IFuncDoubleDoubleComponent>();
             this.line3ComponentTestDouble = new Mock<ILine3Component>();
             this.vector3ComponentTestDouble = new Mock<IVector3Component>();
-            this.systemUnderTest = new ConeSceneComponent(
+            this.systemUnderTest = new CylinderSceneComponent(
                 this.funcDoubleDoubleComponentTestDouble.Object,
                 this.line3ComponentTestDouble.Object,
                 this.vector3ComponentTestDouble.Object);
         }
 
         /// <summary>
-        /// Tests <see cref="ConeSceneComponent.Contains(Vector3)"/>.
+        /// Tests <see cref="CylinderSceneComponent.Contains(Vector3)"/>.
         /// </summary>
         [Fact]
         public void GivenThePointIsNullWhenContainsIsCalledThenAnArgumentNullExceptionIsThrown()
@@ -67,16 +67,16 @@ namespace Scenery.Components.Tests.SceneComponentTests
         }
 
         /// <summary>
-        /// Tests <see cref="ConeSceneComponent.Contains(Vector3)"/>.
+        /// Tests <see cref="CylinderSceneComponent.Contains(Vector3)"/>.
         /// </summary>
         [Fact]
-        public void GivenAPointInsideTheConeWhenContainsIsCalledThenTrueIsReturned()
+        public void GivenAPointInsideTheCylinderWhenContainsIsCalledThenTrueIsReturned()
         {
             // Arrange.
             var point = new Vector3();
             this.vector3ComponentTestDouble
                 .Setup(component => component.DotProduct(It.IsAny<Vector3>(), It.IsAny<Vector3>()))
-                .Returns(-1D);
+                .Returns(0D);
 
             // Act.
             var result = this.systemUnderTest.Contains(point);
@@ -88,16 +88,16 @@ namespace Scenery.Components.Tests.SceneComponentTests
         }
 
         /// <summary>
-        /// Tests <see cref="ConeSceneComponent.Contains(Vector3)"/>.
+        /// Tests <see cref="CylinderSceneComponent.Contains(Vector3)"/>.
         /// </summary>
         [Fact]
-        public void GivenAPointOutsideTheConeWhenContainsIsCalledThenFalseIsReturned()
+        public void GivenAPointOutsideTheCylinderWhenContainsIsCalledThenFalseIsReturned()
         {
             // Arrange.
             var point = new Vector3();
             this.vector3ComponentTestDouble
                 .Setup(component => component.DotProduct(It.IsAny<Vector3>(), It.IsAny<Vector3>()))
-                .Returns(1D);
+                .Returns(2D);
 
             // Act.
             var result = this.systemUnderTest.Contains(point);
@@ -109,7 +109,7 @@ namespace Scenery.Components.Tests.SceneComponentTests
         }
 
         /// <summary>
-        /// Tests <see cref="ConeSceneComponent.GetAllIntercepts(Line3)"/>.
+        /// Tests <see cref="CylinderSceneComponent.GetAllIntercepts(Line3)"/>.
         /// </summary>
         [Fact]
         public void GivenTheLineOfSightIsNullWhenGetAllInterceptsIsCalledThenAnArgumentNullExceptionIsThrown()
@@ -141,7 +141,7 @@ namespace Scenery.Components.Tests.SceneComponentTests
         }
 
         /// <summary>
-        /// Tests <see cref="ConeSceneComponent.GetAllIntercepts(Line3)"/>.
+        /// Tests <see cref="CylinderSceneComponent.GetAllIntercepts(Line3)"/>.
         /// </summary>
         [Fact]
         public void GivenNoInterceptsWhenGetAllInterceptsIsCalledThenNoInterceptIsReturned()
@@ -165,7 +165,7 @@ namespace Scenery.Components.Tests.SceneComponentTests
         }
 
         /// <summary>
-        /// Tests <see cref="ConeSceneComponent.GetAllIntercepts(Line3)"/>.
+        /// Tests <see cref="CylinderSceneComponent.GetAllIntercepts(Line3)"/>.
         /// </summary>
         [Fact]
         public void GivenTwoInterceptsWhenGetAllInterceptsIsCalledThenTwoInterceptsAreReturned()
