+++ SceneryWebApi/Scenery.ComponentsTests/SceneComponents/FullSceneComponentTests.cs
@@ -3,7 +3,7 @@
 // Licensed under the MIT License. See LICENSE file in the project root for full license information.
 // </copyright>
 
-namespace Scenery.Components.Tests.SceneComponentTests
+namespace Scenery.Components.Tests.SceneComponents
 {
     using FluentAssertions;
     using Scenery.Components.Implementations.SceneComponents;
