+++ SceneryWebApi/Scenery.ComponentsTests/SceneComponents/InvertedSceneComponentTests.cs
@@ -3,7 +3,7 @@
 // Licensed under the MIT License. See LICENSE file in the project root for full license information.
 // </copyright>
 
-namespace Scenery.Components.Tests.SceneComponentTests
+namespace Scenery.Components.Tests.SceneComponents
 {
     using System.Collections.Generic;
     using System.Linq;
